initial models
Fixed the DaysScheduledAhead column to be an integer.
Added functions to calculate the F1 score.
Created the logistic regression, ridge, and lasso models and calculated their F1 scores and printed out their confusion matrices.
This is looking an awful lot like HW3.
All that's left is to do elastic net and KNN.
I will work on this more tomorrow night.